+++ src/cli/credrank.js
@@ -9,15 +9,8 @@ import sortBy from "../util/sortBy";
 import {credrank} from "../core/credrank/compute";
 import {CredGraph} from "../core/credrank/credGraph";
 import {LoggingTaskReporter} from "../util/taskReporter";
-import {MarkovProcessGraph} from "../core/credrank/markovProcessGraph";
 import type {Command} from "./command";
 import {loadInstanceConfig, prepareCredData} from "./common";
-import {graphIntervals} from "../core/interval";
-
-const DEFAULT_ALPHA = 0.1;
-const DEFAULT_BETA = 0.4;
-const DEFAULT_GAMMA_FORWARD = 0.1;
-const DEFAULT_GAMMA_BACKWARD = 0.1;
 
 function die(std, message) {
   std.err("fatal: " + message);
@@ -38,30 +31,8 @@ const credrankCommand: Command = async (args, std) => {
   const {weightedGraph, ledger} = await prepareCredData(baseDir, config);
   taskReporter.finish("load data");
 
-  taskReporter.start("create Markov process graph");
-  // TODO: Support loading params from config.
-  const parameters = {
-    beta: DEFAULT_BETA,
-    gammaForward: DEFAULT_GAMMA_FORWARD,
-    gammaBackward: DEFAULT_GAMMA_BACKWARD,
-    alpha: DEFAULT_ALPHA,
-  };
-  const participants = ledger.accounts().map(({identity}) => ({
-    description: identity.name,
-    address: identity.address,
-    id: identity.id,
-  }));
-  const intervals = graphIntervals(weightedGraph.graph);
-  const mpg = MarkovProcessGraph.new({
-    weightedGraph,
-    participants,
-    parameters,
-    intervals,
-  });
-  taskReporter.finish("create Markov process graph");
-
   taskReporter.start("run CredRank");
-  const credGraph = await credrank(mpg);
+  const credGraph = await credrank(weightedGraph, ledger);
   taskReporter.finish("run CredRank");
 
   taskReporter.start("write cred graph");
